+++ apl-conf/src/main/resources/conf-tn1/apl-blockchain.properties
@@ -455,14 +455,14 @@ apl.enableTransactionRebroadcasting=true
 # Set to 0 or negative for no limit.
 apl.maxUnconfirmedTransactions=10000
 
-# Target desirable time of broadcast operation from pending transaction queue in ms. By default set to 1000 ms.
-apl.pendingBroadcast.targetTime=1000
+# Target desirable duration (in ms) of broadcast operation from pending transaction queue in ms. By default set to 1000 ms.
+apl.mempool.pendingBroadcast.targetTime=1000
 
 #  Min number of pending transactions to broadcast in batch by one broadcast iteration. By default is 10
-apl.pendingBroadcast.minBatchSize=10
+apl.mempool.pendingBroadcast.minBatchSize=10
 
 # Max number of pending transactions to broadcast in batch by one broadcast iteration. By default is 255
-apl.pendingBroadcast.maxBatchSize=1024
+apl.mempool.pendingBroadcast.maxBatchSize=1024
 
 # Max size of in-memory transaction queue, which accumulate broadcast-pending transactions. Default is 2000
 apl.mempool.maxPendingTransactions=5000
