+++ apl-conf/src/main/resources/conf-tn3/apl-blockchain.properties
@@ -477,14 +477,14 @@ apl.maxUnconfirmedTransactions=3500
 # Maximum number of account monitors allowed on this node
 apl.maxNumberOfMonitors=100
 
-# Target desirable time of broadcast operation from pending transaction queue in ms. By default set to 1000 ms.
-apl.pendingBroadcast.targetTime=1000
+# Target desirable duration (in ms) of broadcast operation from pending transaction queue in ms. By default set to 1000 ms.
+apl.mempool.pendingBroadcast.targetTime=1000
 
 #  Min number of pending transactions to broadcast in batch by one broadcast iteration. By default is 10
-apl.pendingBroadcast.minBatchSize=10
+apl.mempool.pendingBroadcast.minBatchSize=10
 
 # Max number of pending transactions to broadcast in batch by one broadcast iteration. By default is 255
-apl.pendingBroadcast.maxBatchSize=1024
+apl.mempool.pendingBroadcast.maxBatchSize=1024
 
 # Max size of in-memory transaction queue, which accumulate broadcast-pending transactions. Default is 2000
 apl.mempool.maxPendingTransactions=5000
